Bug: Passport.js does not find the user before attemping to create it
This is the cause of users not being able to login. The email
attribute is required and unique, and since passport.js does not
check if the user already exists (it only checks for a session)
it attempts to create the user and since it's the email already exists
throws an error.